Design change - Don't disable interrupts during direction and interrupt control setup.
After much deliberation, I decided to remove the step which disables all interrupts during an output to input pin change.  Instead, we set the pin direction first, then enable the interrupt.  Before, we were disabling all interrupts reporting, enabling pins for input, then enabling the interrupts.  This behavior was undesirable as it could cause us to miss an interrupt trigger between the time we disabled and re-enabled interrupts should a pin currently configured as an input with interrupts change and trigger an interrupt.